backend/api_fast_swag/requirements.txt
anyio==4.2.0
asgiref==3.7.2
bcrypt==3.2.0
cffi==1.16.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.78.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.35
starlette==0.19.1
typing_extensions==4.9.0
uvicorn==0.17.6
requests
sqlmodel